+++ gaiku-3d/src/formats/gox.rs
@@ -26,9 +26,9 @@ impl FileFormat for GoxReader {
               let block_colors = block_data[data.block_index];
               let mut chunk = Chunk::new([data.x as f32, data.y as f32, data.z as f32], 16, 16, 16);
 
-              for x in 0..chunk.width() {
-                for y in 0..chunk.height() {
-                  for z in 0..chunk.depth() {
+              for x in 0..chunk.width() as usize {
+                for y in 0..chunk.height() as usize {
+                  for z in 0..chunk.depth() as usize {
                     if !block_colors.is_empty(x, y, z) {
                       let color = block_colors.get_pixel(x, y, z);
                       let index = if let Some((index, _)) =
